requirements.txt
discord.py
feedparser
flask
python-dotenv